+++ src/udp.h
@@ -1,7 +1,8 @@
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
 
 #define _1KB_ 1024
-uint8_t* getData(const uint8_t* message, const uint8_t length, int *buf_len);
+uint8_t* getData(const uint8_t* message, const size_t length, int *buf_len);
